requirements.txt
streamlit>=1.32.0
pdfminer.six>=20221105
pdf2image>=1.16.3
pillow>=10.1.0
requests>=2.31.0
groq>=0.3.0
python-dotenv>=1.0.0